sync with latest cowboy code.
Latest cowboy is now using ranch to handle socket connections. This
change reflect it. Also since the cowboy req record is now opaque do a
manual import hoping it won't change often. We could of course use the
cowboy api but it would be came pretty inneficient also it doesn't
expose the buffer.